moving Login/Loading under StartScreen container
