+++ src/containers/StartScreen/Login/Loading/index.js
