+++ src/containers/StartScreen/Login/index.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import Loading from '../Loading';
+import Loading from './Loading';
 
 export default class Login extends Component {
   render() {
